Change profiles directory. Make apps better. Find errors in main/index
