+++ Fitnes/Main/forms.py
@@ -3,7 +3,7 @@ from django import forms
 from django.contrib.auth.models import User
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.forms import AuthenticationForm
-from .models import Profile
+from Profiles.models import Profile
 
 class PasswordResetForm(forms.Form):
     username = forms.CharField(label='Имя пользователя', max_length=150)
@@ -123,28 +123,3 @@ class LoginForm(AuthenticationForm):
         }),
         label=''
     )
-
-class UserProfileForm(forms.ModelForm):
-    email = forms.EmailField(required=True, widget=forms.EmailInput(attrs={
-        'class': 'form-control',
-        'placeholder': 'Почта'
-    }))
-    phone = forms.CharField(max_length=15, required=False, widget=forms.TextInput(attrs={
-        'class': 'form-control',
-        'placeholder': 'Номер телефона'
-    }))
-    username = forms.CharField(max_length=150, widget=forms.TextInput(attrs={
-        'class': 'form-control',
-        'placeholder': 'Имя пользователя'
-    }))
-
-    class Meta:
-        model = User
-        fields = ['username', 'email']
-
-    def save(self, commit=True):
-        user = super().save(commit=commit)
-        if commit:
-            user.profile.phone = self.cleaned_data['phone']
-            user.profile.save()
-        return user
