+++ Fitnes/Main/templates/main/index.html
@@ -41,7 +41,7 @@
   <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js"></script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 
-  <section class="hero" style="background-image: url('{% static "main/img/main.png" %}')">
+  <section class="hero" style="background-image: url('/static/main/img/main.png')">
     <div class="burger-menu">
       <div class="burger-line"></div>
       <div class="burger-line"></div>
@@ -53,7 +53,7 @@
       <a class="dropdown-item" href="/activity/">Активновть</a>
       <a class="dropdown-item" href="/progress/">Достижения</a>
       <a class="dropdown-item" href="{% url 'index' %}">Главная</a>
-      <a class="dropdown-item" href="/fitnes/profile/">Личный кабинет</a>
+      <a class="dropdown-item" href="{% url 'profile' %}">Личный кабинет</a>
       </div>
       <a class="logout" href="/fitnes/login/">Войти</a>
       <div class="hero-text">
@@ -120,7 +120,7 @@
   </div>
   
 
-<section class="workout" style="background-image: url('{% static "main/img/1.png" %}')">
+<section class="workout" style="background-image: url('/static/main/img/1.png')">
     <h2>ПОДДЕРЖИВАЙТЕ СЕБЯ В ФОРМЕ</h2>
     <p>Тренируйте регулярно с помощью новых тренировок, которые можно делать не выходя из дома.</p>
     <div class="workout-content">
@@ -217,42 +217,7 @@
 
 </body>
 </html>
-<!--
-  <div class="container">
-    {% for profile in profiles %}
-    <div class="alert bg-light mb-4">
 
-      <div class="row">
-        <div class="col-md-10">
-          <p class="mb-0">Название профиля: {{ profile.title }}</p>
-        </div> 
-        <div class="col-md-2 text-right">
-          <a class="btn btn-sm btn-info" href="{% url 'task-create' profile_pk=profile.pk %}">Создание задачи</a>
-        </div>
-      </div>
-
-      {% if profile.tasks.all %}
-      <div class="mt-2">
-        {% for task in profile.tasks.all %}
-        <div class="alert bg-secondary text-white d-flex justify-content-between mb-2 py-2 px-3">
-          <div>
-            {{ task.title }} 
-          </div>
-          <div>
-            <a class="btn btn-sm btn-outline-warning" href="{% url 'task-read' task.pk %}">Просмотр</a>
-            <a class="btn btn-sm btn-outline-warning" href="{% url 'task-update' task.pk %}">Редактировать</a>
-            <a class="btn btn-sm btn-outline-danger" href="{% url 'task-delete' task.pk %}">Удалить</a>
-          </div>
-        </div>
-        {% endfor %}
-      </div>
-      {% else %}
-      <p>У этого профиля нет заданий</p>
-      {% endif %}
-    </div>
-    {% endfor %}
-  </div>
-  -->
 
 </body>
 {% endblock content %}
