+++ Fitnes/Profiles/templates/profiles/edit_profile.html
@@ -1,7 +1,7 @@
 {% extends "main/base.html" %}
 
 {% block content %}
-<h1>Редактирование профиля</h1>
+<h1>Редактирование профиля калорий</h1>
 <form method="post">
     {% csrf_token %}
     {{ form.as_p }}
